nhwc convolutions are working perfectly, checked with the test code. The nchw convolutions are also very similar and have almost same values, but i should work on reducing that gap.
